split async and sync tests to avoid sync and async event-loops which async_alru does not like
