+++ setup.cfg
@@ -53,7 +53,7 @@ console_scripts =
 [tox:tox]
 labels =
     test = py310, py311, py312
-    lint = mpy 
+    lint = mpy
 min_version = 4.0
 env_list =
     mpy
@@ -74,7 +74,9 @@ deps =
     pytest-dependency
     pytest-mock
 commands =
-    python -m pytest --doctest-modules . 
+    python -m pytest --doctest-modules tests
+    python -m pytest --doctest-modules tests_async
+
 [mpy]
 commands_pre =
     python --version
