+++ src/aerovaldb/lock/lock.py
@@ -89,7 +89,7 @@ class FileLock(AerovaldbLock):
         logger.debug("Acquiring lock with lockfile %s", self._lock_file)
 
         if has_async_loop():
-            run_until_finished(self._aiolock.acquire)
+            run_until_finished(self._aiolock.acquire())
 
         fcntl.lockf(self._lock_handle, fcntl.LOCK_EX)
         self._acquired = True
